Add files via upload
Excahnge the 0.0.0.0 in the Forward Lookup Zon with your IP address of the client and in the Revers file the 0 with the last octet of your IP address. e.g. with the IP address 192.168.178.23 you would exchange the zero for the 23. And the ending .github must also removed from the files.